review servlet needed to set REDIRECT_TO attribute in session
welcome page extensions have an in-built redirect mechanism based on the
return message from the login XHR. the login command on the server looks
for the REDIRECT_TO attribute in the session and returns the value of
that attribute -- assumed to be a valid URL -- as the response to the
login XHR. the welcome page login logic looks at the XHR result. if the
result is the string "OK", the login logic stays on the same page. if
the result string is anything else the login logic sets 'location.href'
to the result string.