Enhancement:
(1) All annotations for the current state are (by default) always visible, but if any comments are selected, then annotations for unselected comments are partially grayed out, probably using opacity would be ~0.5 (not sure of the best value).
(2) We didn't talk about this during the phone call, but I would eliminate the "nothing selected/comment open" mode, which is the behavior Janet is describing in #30 here. When a comment is open, then that comment should be selected upon opening and all other comments should be unselected. When combined with (1), that would mean that all annotations corresponding to the currently open comment would be fully visible and all other annotations would be partially grayed out.