temporary fix 901
save and saveAs functions should use same code path. currently, save
function calls getText() on editor whereas saveAs function (in a
completely different class) calls getText() on the model. the correct
thing to do is to call getText() on the model everywhere. this change is
temporary until every flavor of save and saveAs can be unified to
getText() from the model.