Set package locations for all user libs on page creation.
For now, create a 'package' declaration in data-dojo-config
for all the user libs.  Makes it so we don't have to set the
module path dynamically in the library code.
In the future, we may want to make this work dynamically, such
that the 'package' decl is added when a widget of that lib
is first added to the page.