Add support for getTargetOverlays helper function (#1732)
The bulk of this work was to modify _Tool.js so that it would respect a
new getTargetOverlays helper function. And, when present, use the
rectangular region(s) returned by the helper rather than a single
rectangular region calculated from the boundaries of the widget in
question. The core logic of _Tool.js was maintained, but functions were
added to handle showing/hiding the target overlay(s) to abstract the
use of an array of overlays.

The rational for adding this functionality is that certain widgets
(like HorizontalSlider) are treated as a single control in the
Visual Editor. But, the user also needs to be able to individually
select its children to further customize (e.g., in case of
HorizontalSlide these children are HorizontalRule and
HorizontalRuleLabels). So, we wanted to create a getTargetOverlays
helper so a set of rectangles that will leave these children uncovered
by the can be calculated.

For the time being, this support is being added primarily for Horizontal
and Vertical Sliders. But, other widgets may leverage this function
moving forward.