Make unload handler in CommentView work in FF (#1867)
Before diving into the core of #1867, it was discovered that the
beforeunload handler in CommentView was not causing a dialog to show up
in FF when running the scenario originally described in the bug report.
So, added the appropriate work around for FF. Also, instead of using
dojo.connect switched to use dojo.addUnload. Finally, tweaked Runtime.js
mostly to make use of dojo.addUnload there too.