Close holes when user tries to leave Maqetta with unsaved review comments (#1867)
This commit includes a number of changes to try and make sure the user
is always warned if they try to navigate away from Maqetta with unsaved
review comments. Changes were made to the unload handler in
Workbench.js to give each open editor a chance to specify a
warning message. This is a general feature any editor could us, but
ReviewEditor is the only one does because it's a
special case in how it handles comments modally.

A number of other changes were made as part of this work: Comments tab
cleared when user switches to a non-review editor, more reliable
tracking of isDirty flag for review editors, marking tab title of dirty
review with an "*" like any other dirty editor, remove unload handler in
CommentView.js, etc.