Fix inability for user to switch state in theme editor (#1894)
StatesView contained a _silent flag which was causing the event for the
first user-inititated state change to not be fired. Perhaps the flag
once served a purpose, but the only place it was being consulted was in
the "onclick" handler for the states palette. And, it really doesn't
make sense to ever ignore the user-initiated click. So, removing all
code related to the _silent flag in StatesView.