Review comment replies overlapping other comments (#1952)
The root of this problem is that the commentReplies class defined in
Comment.css hardcoded a height of 7px. So, when a reply would be
expanded, the div holding the reply was not allowed to get larger to
accommodate the content of the reply. So, the conent of the reply went
past of the edges of the div and overlapped other comments. So, I
removed the height restriction.

In addition, after removing the height restriction I'd sometimes see a
partial "X" icon or a partial "X" with glasses icon outside of the
reply div. The commentReplies class set a background image based on
sprite.png, but this seemed to serve no useful purpose. So, I removed
that
as
well.