First wave of changes so user can see both reviews he's created and reviews he's been invited to in the same tree (#360)
This commit includes a variety of server and client changes. At a
high-level...

On the server, changes were made to maintain an index for
each reviewer that includes all reviews they've been invited to (which
includes reviews created by the reviewers himself). This necessitated 
changes to necessary REST commands (including honoring a designerId
passed in as an arg, etc.).

On the client, changes were made to account for reviewers having
individual designers, passing additional args back to the server, etc.

There are still some open issues on both client and server, and most of
those are tagged in code with "AWE TODO:". I'll continue to work through
those items.