Added GridX configuration wizard (part of #331)
Added a wizard to configure and instance of GridX consisting of 3
panels: 1) data source configuration, 2) column selection, 3) preview &
column re-sizing/re-ordering. This replaces the GridX smart input that
had been enabled for in an earlier sprint of M6.

Most of the code updates were within the gridx plugin. However, a
number of changes were made within the dojo 1.7 plug-in (mostly
DataStoreBasedWidgetInput and DataGridInput) so that as much as possible
of the existing Smart Input support for Data Grid and GridX could be
reused an an embedded div within the first wizard panel. The goal was to
leave Data Grid input working exactly as it was before while enabling
the GridX wizard.