Fix link in review invitation (#2108)
The DavinciReviewServlet was recently disabled. So, review link URL's
now need to be processed by DavinciPageServlet. This means the old
review link URL's using /review as the context root don't make sense
anymore.

Added processing of the review-related attributes to
DavincePageServlet. And, along with removing /review from the URLs,
took the opportunity to further simplify the URL and make it easier to
read. 

Links to reviews will now be in the form of:

http://localhost:8081/maqetta?reviewDesigner=userA&reviewVersion=20120502112254