Fix JS error introduced in b51206cdd92f10b0534c213e58ba5109be4cc8ea for non-HTML editors (#1435)
It was pointed out to me that the change for
b51206cdd92f10b0534c213e58ba5109be4cc8ea causes a JS error when
selecting text in editors for non-HTML files types (e.g.,
"this.model.mapPositions is not a function"). So, I added a base
implementation of mapPositions to Model.js that returns unchanged
indices.