Fix unsanitized cookie for mobile preview (Bugzilla 7973)
Fixed this time by just removing the code that adds the
"preview" cookie. I searched high and low for any other code that looks
for that cookie, and could not find any other reference to it. The most
likely candidate would have been preview.html, but it gets
the value out that parm out of the URL.