Revert "Make it so that Joomla's welcome.html page redirects to Joomla login"
The fix didn't work. @childbs and @JonFerraiolo discussed and concluded that
best to leave this unchanged versus previous releases. It's not that bad of
a usability thing that maqetta.org users logoff and are taken to home page.
Also, @childsb and @JonFerraiolo feel that the whole login/logout mechanism
might change at Maqetta.org once Orion integration is complete, and Joomla
might be totally out of loop. Therefore, it seemed prudent to avoid risk
for M6 and leave things as they were.

This reverts commit db85d7f07951f83101cf736385ec822f2b96b7a0.