Add capability for UI widget plugins to add context menu actions (#2129)
As the first step of the work for #2129 to better support HTML table
column editing (resizing, etc.) added capability to look at package.json
metadata to add action sets.

As part of that work, moved all of the table related actions out of
the core.client plugin and into the davinci.html plugin.