Fix undo/redo with add columns/rows (#2566)
The main fix here wasn't really HTML table-specific, but was correcting
a small logic error in AddCommand which prevented undo from working when
a piece of widget data was passed instead of an actual widget instance.

In addition, a try/catch block was added to VisualOutlineEditor because
it seems to have occasional failures in the _widgetChanged function.
These failures aren't table specific and aren't necessarily undo/redo
related, but they can interfere with undo/redo. A separate issue will be
opened to investigate the outline issues.