Incorrect initial GridX display when using JSON file (#2575)
The cause of the symptom described in #2575 was a one-line change in
commit 6836ed983f3a805fab250dd63c9e1994eae0a11 . However, as that change
was backed out, it was disovered there were general problems switching
between data source types and file types (particularly when CSV files
enter the picture). So, more widespread changes were made to try and
stabilize this area of code.

In addition, the GridX test cases were beefed up in a few ways, but most
notably by adding a test case for CSV files. In addition, a test
document was added for `EdgeToEdgeDataList` and `Carousel` mobile
widgets they depend on `DataStoreBasedWidgetInput` and we need a better
way to find potential regression problems here.
because
they depend on DataS