Add ignore=dirty to all submodules in .gitmodules
We're getting a lot of submodules, and it's getting annoying to
see them appear in the modified list when we've not changed the HEAD
they point too. Since we're not planning to edit source within our
submodules (treating them as read-only), I've added ignore=dirty to all
of them in our .gitmodules file. This means the following (according to
the gitmodules docs):

> will ignore all changes to the submodules work tree and takes only
differences between the HEAD of the submodule and the commit recorded in
the superproject into account.

Unfortunately, even with this change EGit still seems to list the
submodules as candidates to be committed. Perhaps this will be fixed in
later versions of EGit.