Review publish not working (#2658)
The publish operation in PublishWizard.js uses dojo.xhrPost, and on the
server side getParameter in Publish.java was returning null for all
parameters. Presumably, something changed in the dojo.xhrPost code
for Dojo 1.8 because I'm not seeing any other recent changes in either
PublishWizard.js or Publish.java. So, I changed to use
dojo.objectToQuery() to build the query parms as alluded to
in the docs
(http://livedocs.dojotoolkit.org/dojo/xhrPost#example-1-dojo-xhrpost-call-to-send-a-form):
"If you want to send some parameters in a query string, while making
POST, you should include them in url yourself. Dojo provides a special
helper for that: dojo.objectToQuery()"

Generally, we don't use xhrPost very much (even though we have a number
of URI's that create objects on the server and should probably
technically be POST operations). I only found one other instance of
dojo.xhrPost (and it was exhibiting the same sort of problem). That
xhrPost was used for updating comments in Comment.js. Since
the add comment operation (in the same Comment.js file) uses xhrGet, I
just changed to xhrGet for updating comments as well.

Unfortunately, in fixing the issue in Publish.js, I uncovered an new
issue when running from the local Eclipse environment.
SmtpPop3Mailer.java threw a NullPointerException trying to figure out
the mail server. So, I fixed that.

And, unfortunately, this uncovered yet another issue where the dialog
that shows the user the message that would have been in e-mail just said
"E-mail server could not be found". So, cleaned up the response code
on the server side to use the html content like it did before.