Tolerate "legacy" tree setup (#2687)
Added code to ensure we could edit trees that were setup in the way we
did before the addition of tree smart input (#79). E.g., using
ItemFileReadStore and ForestTreeModel.

The general approach employed in TreeInput.js to accomplish this is to
go into a sort of "legacy" mode if ItemFileReadStore is encountered.
The data from the
ItemFielReadStore is loaded and converted to the relational format
desired by Memory. This converted data is then passed into the Memory
object that backs the preview. 

When the user clicks OK, we get the data out of the Memory object and
convert it back to the format expected by ItemFileReadStore. This allows
us to reuse all of the UI that had been built up for tree manipulation
in the smart input dialog while continuing to maintain ItemFileReadStore
in the designer's page.

Also, added a short section to the tree's test document.