(Orion) R&C Can't create reviews (looks like bad command paths) [#2950]
R&C JS code was (unnecessarily) building up URLs with code like:

var location = Workbench.location().match(/http:\/\/.*:\d+\//);
var latestVersionId = Runtime.serverJSONRequest({
	url: location + "maqetta/cmd/getLatestVersionId",
	sync: true
}); 

On hosted orion, the location was coming back as "null" because there's
no port on the URL, but the regular expression used in the call to
`match` assumes there's a port. This caused a bad URL to be formed.

In the end, it's sufficient to do the following:

```
url: "cmd/getLatestVersionId",
```