Create "maq-lib-<libName>" AMD module namespace.
This is necessary for use by workbench code which wishes
to use the library code.  For example, the GridX wizard creates
an example GridX instance in the workbench context (not VE),
using the GridX code from the widget library.
For each library, create the AMD module and a "packageMap",
which correctly maps required any required resources to the
correct AMD module (i.e. the GridX code loaded from "maq-lib-gridx"
makes use of Dojo, which should be loaded from "maq-lib-dojo";
ensures that Dojo is loaded from the widget plugin and doesn't
use the workbench's Dojo, which may be different versions.

Build the GridX code without explicit module IDs (set
"insertAbsMids" to 0 in the build profile).  This ensures that
the GridX code is "relocatable" (i.e. can use an AMD module
other than "grids").