Remove "Exporting projects to IBM Rational Application Developer (TM) 8.5 without including the Dojo toolkit" section (#3266)
I had updated the "Exporting projects to IBM Rational Application
Developer (TM) 8.5 without including the Dojo toolkit" section in the
Maqetta docs on exporting from Maqetta in commit
4ced05c104ea0be783144521866fc4e8fd5d68db for #3266. 

However, I then decided the section should really not even exist for M7
because RAD 8.5 is still on Dojo 1.7. Since Maqetta is now on Dojo 1.8,
I think folks are asking for trouble trying to follow these instructions
to wind up with a RAD workspace with Dojo 1.7 in place. For example, I
tried with something really simple that you'd think you work (like a
Dijit Calendar on a page), but then the rendering was wrong because of
mismatched theming.

Perhaps this could be added back in the future when RAD is back on the
same level of Dojo. I'll open a new issue for that.