Differentiate between a user "login" and "id".
Most of the UserManager code expects a user "id", in the form
of "A" or "bt".  The reset password logic doesn't have that --
it only has an email, a user "login".
Too late in M7 to come up with a proper fix, so instead we
add support for both paths.
#3350