Fix ordering of data store relative to 2nd data store based widget (#3254)
@doronrosenberg added commit 0aa4b4c445584754292cd51e3536a6f632d6768f to
stop reparenting during the create function in DataGridHelper to fix
#3345. That commit fixes the core problem in #3254 as well (e.g., the
2nd data store coming after the 2nd GridX and subsequent RSOD).

But, with 0aa4b4c445584754292cd51e3536a6f632d6768f the data store for
the 2nd GridX gets put at the very top when Finish is clicked on the
GridX wizard (rather than right above it). So, this new commit is in
place to correct that now that reparent no longer occurs during create.