Annotations not displayed correctly - user compare problem (#3355)
In certain scenarios, user id tends to be unreliable in R&C for
reviewers. For example, after the server is restarted, user ids become
null until the review has logged in again (see #3407). And, there are
some other (seemingly less reproducible situations) where the id seems
to change (#3406). So, the modifications for this commit, involve
relying on e-mail when determining colors for annotations/comments, what
annotations to display, etc.