Don't return failed req for AddComment w/ no email server.
Make AddComment work more like Publish: if server cannot
reach email server or it is not configured, still handle
the request as successful (since server was able to add
comment or publish new review session).  Return the email
error as part of result and display warning message to user
in the client.
Fixes #3915